+++ common/src/main/java/me.wiefferink.errorsink.common/CommonClass.java
@@ -1,0 +1,5 @@
+package me.wiefferink.errorsink.common;
+
+public class CommonClass {
+
+}
