+++ pom.xml
@@ -9,6 +9,14 @@
     <version>1.2.0-SNAPSHOT</version>
     <name>ErrorSink</name>
 
+    <modules>
+        <module>common</module>
+        <module>spigot</module>
+        <module>sponge</module>
+    </modules>
+
+    <packaging>pom</packaging>
+
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <maven.compiler.source>1.8</maven.compiler.source>
@@ -16,195 +24,11 @@
     </properties>
 
     <repositories>
-        <repository>
-            <id>spigot-repo</id>
-            <url>https://hub.spigotmc.org/nexus/content/repositories/snapshots/</url>
-        </repository>
+
         <repository>
             <id>bstats-repo</id>
             <url>http://repo.bstats.org/content/repositories/releases/</url>
         </repository>
     </repositories>
-
-    <dependencies>
-
-        <!-- Bukkit API -->
-        <dependency>
-            <groupId>org.bukkit</groupId>
-            <artifactId>bukkit</artifactId>
-            <version>1.8-R0.1-SNAPSHOT</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Sentry-Java log4j2 bindings to send events to Sentry.io -->
-        <dependency>
-            <groupId>io.sentry</groupId>
-            <artifactId>sentry-log4j2</artifactId>
-            <version>1.5.2</version>
-        </dependency>
-
-        <!-- Prevent Raven-Java warning about logging framework -->
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-nop</artifactId>
-            <version>1.7.21</version>
-        </dependency>
-
-        <!-- Bstats -->
-        <dependency>
-            <groupId>org.bstats</groupId>
-            <artifactId>bstats-bukkit</artifactId>
-            <version>1.1</version>
-        </dependency>
-
-    </dependencies>
-
-    <profiles>
-        <!-- Copy the jar file to a test server, activate by using -DcopyResult="/path/to/test/server" -->
-        <profile>
-            <id>test-locally</id>
-            <activation>
-                <property>
-                    <name>copyResult</name>
-                </property>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <artifactId>maven-antrun-plugin</artifactId>
-                        <version>1.8</version>
-                        <configuration>
-                            <target>
-                                <copy file="./target/${project.build.finalName}.jar"
-                                      tofile="${copyResult}\${project.build.finalName}.jar"/>
-                            </target>
-                        </configuration>
-                        <executions>
-                            <execution>
-                                <phase>package</phase>
-                                <goals>
-                                    <goal>run</goal>
-                                </goals>
-                            </execution>
-                        </executions>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-
-        <!-- Package the source files in a jar, activate by using -Dsources -->
-        <profile>
-            <id>package-sources</id>
-            <activation>
-                <property>
-                    <name>sources</name>
-                </property>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-source-plugin</artifactId>
-                        <version>3.0.0</version>
-                        <executions>
-                            <execution>
-                                <phase>package</phase>
-                                <id>attach-sources</id>
-                                <goals>
-                                    <goal>jar</goal>
-                                </goals>
-                            </execution>
-                        </executions>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-
-        <!-- Produce javadoc files, activate by using -Djavadoc -->
-        <profile>
-            <id>generate-javadoc</id>
-            <activation>
-                <property>
-                    <name>javadoc</name>
-                </property>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-javadoc-plugin</artifactId>
-                        <version>2.10.4</version>
-                        <executions>
-                            <execution>
-                                <phase>package</phase>
-                                <id>attach-javadocs</id>
-                                <goals>
-                                    <goal>jar</goal>
-                                </goals>
-                            </execution>
-                        </executions>
-                        <configuration>
-                            <destDir>${javadoc}</destDir>
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-    </profiles>
-
-    <build>
-        <directory>target</directory>
-        <finalName>ErrorSink</finalName>
-        <resources>
-            <resource>
-                <targetPath>.</targetPath>
-                <filtering>true</filtering>
-                <directory>src/main/resources</directory>
-                <includes>
-                    <include>plugin.yml</include>
-                    <include>config.yml</include>
-                </includes>
-            </resource>
-        </resources>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-shade-plugin</artifactId>
-                <version>2.4.3</version>
-                <executions>
-                    <execution>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>shade</goal>
-                        </goals>
-                    </execution>
-                </executions>
-                <configuration>
-                    <relocations>
-                        <!-- Relocate Sentry -->
-                        <relocation>
-                            <pattern>io.sentry</pattern>
-                            <shadedPattern>me.wiefferink.errorsink.shaded.sentry</shadedPattern>
-                        </relocation>
-                        <!-- Relocate bstats -->
-                        <relocation>
-                            <pattern>org.bstats</pattern>
-                            <shadedPattern>me.wiefferink.errorsink.shaded.bstats</shadedPattern>
-                        </relocation>
-                    </relocations>
-                    <artifactSet>
-                        <excludes>
-                            <!-- Already in spigot -->
-                            <exclude>org.apache.logging.log4j</exclude>
-                        </excludes>
-                    </artifactSet>
-                </configuration>
-            </plugin>
-            <plugin>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.5.1</version>
-            </plugin>
-        </plugins>
-    </build>
 
 </project>
