--- src/main/resources/plugin.yml
@@ -1,19 +1,0 @@
-name: ErrorSink
-description: "Collect server errors and warnings in a beautiful web interface"
-main: me.wiefferink.errorsink.ErrorSink
-version: "${project.version}"
-author: NLThijs48
-website: http://github.com/NLthijs48/ErrorSink
-load: STARTUP
-# Ask to load before commonly used plugin to try loading as first and disabling last
-loadbefore:
-  - Vault
-  - WorldEdit
-  - ProtocolLib
-  - LuckPerms
-  - AsyncWorldEditInjector
-  - MinecraftMarket
-  - OpenInv
-  - dynmap
-  - OreBroadcast
-  - GoldRush
